Se ha sacado la lógica del endpoint para meterlo en un componente
